# openapi.yaml
openapi: 3.0.3
info:
  title: Ordryn / GoTodo API
  description: |
    JSON API for Ordryn (GoTodo). Self-host with `GOTODO_MODE=full` (UI + API) or
    `GOTODO_MODE=api` (API only). Redis is required for `/api/v2` authentication,
    rate limiting, and live SSE fan-out.

    Auth: session cookie (SPA login/register) or `Authorization: Bearer <api_key>`.

    Rate limits (per user, Redis token bucket). Reads and writes are independent.
    Bearer API keys: 120 GET/min (~2/sec) and 60 writes/min (~1/sec).
    SPA session cookies use a higher budget (600 GET/min, 240 writes/min) so the
    web UI is not gated by the external API ceiling.

    Canonical paths are `/api/v2`. `/api/v1` remains as a compatibility alias
    for existing webhooks, OAuth callbacks, and clients.
  version: 2.0.0
  license:
    name: See LICENSE in repository
servers:
  - url: /
    description: Same-origin (BASE_PATH may prefix these paths)
...
paths:
...
  /api/v2/me/extensions:
    get:
      tags: [Extensions]
      summary: List personal-inbox hook destinations
      description: Only site-enabled extensions that declare member-scoped settings are returned.
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Personal destinations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectExtensionsList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v2/me/extensions/{extensionId}:
    parameters:
      - name: extensionId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Extensions]
      summary: Get a personal-inbox destination
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Personal destination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectExtension"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Extensions]
      summary: Update a personal-inbox destination
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectExtensionPatch"
      responses:
        "200":
          description: Updated destination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectExtension"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v2/projects/{id}/extensions:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Projects]
      summary: List site-enabled project-scoped extensions for members
      description: Only extensions enabled in Admin → Extensions and that declare a project surface are returned.
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Site-enabled extensions with this project's settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectExtensionsList"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v2/projects/{id}/extensions/{extensionId}:
    parameters:
      - $ref: "#/components/parameters/ID"
      - name: extensionId
        in: path
        required: true
        schema:
          type: string
    patch:
      tags: [Projects]
      summary: Update this project's extension settings (owner)
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectExtensionPatch"
      responses:
        "200":
          description: Updated project extension
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectExtension"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v2/projects/{id}/invites:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Projects]
      summary: List pending project invites
      description: Any project member can list pending invites. Creating and revoking invites remains owner-only.
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Invites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectInvite"
    post:
      tags: [Projects]
      summary: Invite a user by username (owner only)
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInviteCreate"
      responses:
        "200":
          description: Invite created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectInvite"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Project not found, or username not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v2/projects/{id}/invites/{inviteId}:
    parameters:
      - $ref: "#/components/parameters/ID"
      - name: inviteId
        in: path
        required: true
        schema:
          type: integer
    delete:
      tags: [Projects]
      summary: Revoke a pending invite (owner only)
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "204":
          description: Revoked
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v2/invites:
    get:
      tags: [Invites]
      summary: List invites
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Invites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invite"
    post:
      tags: [Invites]
      summary: Create an invite
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                expires_at:
                  type: string
                  format: date-time
                bypass_expiration:
                  type: boolean
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
  /api/v2/invites/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    delete:
      tags: [Invites]
      summary: Delete an unused invite
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "204":
          description: Deleted
...
  /api/v2/admin/extensions:
    get:
      tags: [Admin]
      summary: List loaded extensions
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Extensions with site-level enable state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminExtensionsList"
  /api/v2/admin/extensions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Admin]
      summary: Get one extension and its settings
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Extension
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminExtension"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Admin]
      summary: Enable or disable an extension for the site
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminExtensionPatch"
      responses:
        "200":
          description: Updated extension
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminExtension"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v2/admin/invites:
    get:
      tags: [Admin]
      summary: List all invites (admin)
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: All invites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invite"
    post:
      tags: [Admin]
      summary: Create an invite (admin)
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                expires_at:
                  type: string
                  format: date-time
                bypass_expiration:
                  type: boolean
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
  /api/v2/admin/invites/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    delete:
      tags: [Admin]
      summary: Delete an unused invite (admin)
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "204":
          description: Deleted
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: API key from profile, bootstrap, or device auth
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: httpOnly session cookie from /api/v2/auth/login or register
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable code (e.g. invalid_request)
        message:
          type: string
...
    ProjectInvite:
      type: object
      properties:
        id:
          type: integer
        project_id:
          type: integer
        email:
          type: string
        user_name:
          type: string
        role:
          type: string
          enum: [editor, viewer]
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        project_name:
          type: string
        inviter_email:
          type: string
        inviter_user_name:
          type: string
    ProjectInviteCreate:
      type: object
      required: [username, role]
      properties:
        username:
          type: string
          description: Unique username of the user to invite
          minLength: 3
          maxLength: 32
        role:
          type: string
          enum: [editor, viewer]
...
    Invite:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        token:
          type: string
        used:
          type: boolean
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_by:
          type: integer
          nullable: true
        creator_user_name:
          type: string
        creator_email:
          type: string
        status:
          type: string
          enum: [pending, used, expired]
...
    AdminExtensionSettings:
      type: object
      properties:
        enabled:
          type: boolean
        triggers:
          type: array
          items:
            type: string
        templates:
          type: object
          additionalProperties:
            type: string
        last_error:
          type: string
        last_delivery_at:
          type: string
    AdminExtension:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        host_api:
          type: integer
        status:
          type: string
          enum: [loaded, failed]
        error:
          type: string
        manifest:
          type: object
          additionalProperties: true
        settings:
          $ref: "#/components/schemas/AdminExtensionSettings"
        secrets:
          type: object
          additionalProperties:
            type: boolean
        deliveries:
          type: array
          items:
            $ref: "#/components/schemas/ExtensionDelivery"
    AdminExtensionsList:
      type: object
      properties:
        extensions:
          type: array
          items:
            $ref: "#/components/schemas/AdminExtension"
    AdminExtensionPatch:
      type: object
      properties:
        enabled:
          type: boolean
        webhook_url:
          type: string
        triggers:
          type: array
          items:
            type: string
        templates:
          type: object
          additionalProperties:
            type: string
    ProjectExtensionSettings:
      type: object
      properties:
        enabled:
          type: boolean
        triggers:
          type: array
          items:
            type: string
        templates:
          type: object
          additionalProperties:
            type: string
        status_only:
          type: boolean
        skip_self:
          type: boolean
        min_priority:
          type: integer
        tag_ids:
          type: array
          items:
            type: integer
        status_ids:
          type: array
          items:
            type: integer
        status_exclude_ids:
          type: array
          items:
            type: integer
        claimed_only:
          type: boolean
        claimed_is_me:
          type: boolean
        field_key:
          type: string
        field_value:
          type: string
        quiet_hours_start:
          type: string
        quiet_hours_end:
          type: string
        digest:
          type: string
        mention_map:
          type: object
          additionalProperties:
            type: string
        values:
          type: object
          additionalProperties:
            type: string
        last_error:
          type: string
        last_delivery_at:
          type: string
    ProjectExtension:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        host_api:
          type: integer
        site_enabled:
          type: boolean
        manifest:
          type: object
          additionalProperties: true
        settings:
          $ref: "#/components/schemas/ProjectExtensionSettings"
        secrets:
          type: object
          additionalProperties:
            type: boolean
        member:
          $ref: "#/components/schemas/ProjectExtensionSettings"
        member_secrets:
          type: object
          additionalProperties:
            type: boolean
        signing_set:
          type: boolean
        member_signing_set:
          type: boolean
        signing_secret:
          type: string
          description: Plaintext HMAC secret, shown only immediately after rotate_signing.
        sample_json:
          type: string
        deliveries:
          type: array
          items:
            $ref: "#/components/schemas/ExtensionDelivery"
        member_deliveries:
          type: array
          items:
            $ref: "#/components/schemas/ExtensionDelivery"
    ProjectExtensionsList:
      type: object
      properties:
        extensions:
          type: array
          items:
            $ref: "#/components/schemas/ProjectExtension"
        is_owner:
          type: boolean
    ProjectExtensionPatch:
      type: object
      properties:
        enabled:
          type: boolean
        triggers:
          type: array
          items:
            type: string
        templates:
          type: object
          additionalProperties:
            type: string
        status_only:
          type: boolean
        skip_self:
          type: boolean
        min_priority:
          type: integer
        tag_ids:
          type: array
          items:
            type: integer
        status_ids:
          type: array
          items:
            type: integer
        status_exclude_ids:
          type: array
          items:
            type: integer
        claimed_only:
          type: boolean
        claimed_is_me:
          type: boolean
        field_key:
          type: string
        field_value:
          type: string
        quiet_hours_start:
          type: string
        quiet_hours_end:
          type: string
        digest:
          type: string
          enum: ["", hourly, daily]
        mention_map:
          type: object
          additionalProperties:
            type: string
        webhook_url:
          type: string
          description: Write-only incoming webhook URL. Omit or blank to keep the stored secret. Allowed hosts depend on the extension delivery type (Discord, Slack, Teams, ntfy, or generic HTTPS).
        ntfy_auth:
          type: string
        rotate_signing:
          type: boolean
        rotate_callback:
          type: boolean
        values:
          type: object
          additionalProperties:
            type: string
    ExtensionDelivery:
      type: object
      properties:
        id:
          type: integer
          format: int64
        event:
          type: string
        event_id:
          type: string
        host:
          type: string
        status:
          type: string
        http_code:
          type: integer
        error:
          type: string
        attempts:
          type: integer
        created_at:
          type: string
          format: date-time
...